Incorporated parall. funs of assig. 2 and compared aggainst hpc_tutor seq. funs
I took the functions from assignment 2
and moved them into a separate linalg_t file.
That way, they become part of the library and can be used in future assignments.
The sequential versions were mostly reimplementations of things we had.
When they were, I deleted them and compared the parallel version to our version.
When we did not have them, I added them to the linalg file.
Now we have search in and sort of vectors.